+++ taxsystem/urls.py
@@ -54,6 +54,11 @@ urlpatterns = [
         views.update_tax_period,
         name="update_tax_period",
     ),
+    path(
+        "corporation/<int:corporation_id>/manage/member/<int:member_pk>/delete/",
+        views.delete_user,
+        name="delete_user",
+    ),
     # -- Tax Payment System
     path(
         "corporation/<int:corporation_id>/manage/user/<int:user_pk>/switch_user/",