+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - Update Account on Corporation Change
   - Reactivate Account on returning to Corporation
 - `TAXSYSTEM_STALE_TIME` Defines the time (in days) after which data is considered outdated and needs a update
+- Member Delete Button
+  - Delete Missing Member from Corp Members list
 
 ### Changed